requirements.txt
flask
fastapi
uvicorn[standard]
opentelemetry-proto

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
